Fix ActiveRecord 4.2 support with backport of Relation#in_batches
The ActiveRecord::Relation#in_batches method allows to access already batched relation objects.

For us it allows to fetch record attributes without instantiating AR model objects which is slow and heavy.

Unfortunately it was introduced only in ActiveRecord 5.0.

To prevent breakage of anything we backporting it with refinements (so it is only available inside RelationDumper class).