Add ability to ignore columns from a model (#17)
When dumping a specific model we might want to skip some columns in the
dump. This is specially helpful when those columns are encrypted and we
don't have access to the keys to decrypt them.

We need to to this at the `dump!` level (and not as a
customizer/anonymizer) because we don't want Rails to even query for the
column without the decrypt keys (since that will raise an error).